starting to play with creating classes
