+++ lib/csv_loader.rb
@@ -1,0 +1,21 @@
+require 'csv'
+
+class CsvLoader
+
+  attr_reader :games, :teams, :game_teams
+
+  def initialize (games, teams, game_teams)
+    @games = games
+    @teams = teams
+    @game_teams = game_teams
+  end
+
+
+  # def self.read_csv(locations)
+  #   games = CSV.table(locations[:games])
+  #   teams = CSV.table(locations[:teams])
+  #   game_teams = CSV.table(locations[:game_teams])
+  #   CsvLoader.new(games, teams, game_teams)
+  # end
+
+end
